Add support for extra publication metadata.
In particular this change adds support for OSSRH required metadata for
publishing to maven central.  An object is exposed for pantsbuild/pants
use that encapsulates the boilerplate common to all jars it will publish
to maven central and this object is used to prepare all publishable java
targets for maven central publishing.