Example configuration for the isolated strategy.
- Example configuration for the isolated strategy.

This is approximately the configuration we use with the isolated strategy and zinc for java. We may not want to actually commit this upstream, but it might be useful to have for posterity.

The last missing piece of our effort was a quick one-off Task that jars classpath entries that match a particular regex. This was necessary because as discussed previously, giving the JVM 1100 input directories (900 targets + 200 resource targets) did not work out terribly well. We'll be working to upstream better solutions to that problem soon.

Reviewed at https://rbcommons.com/s/twitter/r/2185/