Use the default ivysettings.xml provided by ivy.
The pantsbuild/pants repo no longer has any dependencies in the
maven.twttr.com repo and so can just use the default ivysettings.xml
embedded in the ivy jar itself for all resolution.

This required replacement of the custom libthrift hosted at
maven.twttr.com with an un-optimized/un-bugfixed one published by
hadoop.

This includes a breaking change to installations relying on the default
location of a custom ivysettings.xml. A placeholder note is added in the
CHANGELOG to save space for warning users.