Move a comment to work around a pytest bug.
When that test fails, pytest attempts to construct a verbose backtrace.
However that comment in the middle of a multiline list literal tickles
a bug that causes pytest to fail hard with an INTERNALERROR, instead
of showing the details of the underlying bug.

While debugging this, I noticed that we don't constrain the versions of
pytest that we bake into the test chroot, so it will end up being a
different version than the one in 3rdparty/python/requirements.txt.
This is pretty confusing, so I added the relevant constraints.