Move a comment to work around a pytest bug.
When that test fails, pytest attempts to construct a verbose backtrace.
However that comment in the middle of a multiline list literal tickles
a bug in the backtrace construction code that causes pytest to fail hard
with an INTERNALERROR, instead of showing the details of the underlying bug.

While debugging this, I noticed that we don't constrain the versions of
pytest that we bake into the test chroot, so it will end up being a
different version than the one in 3rdparty/python/requirements.txt.
This is pretty confusing, so I added the relevant constraints.

See here for an example of this useless INTERNALERROR dump:
https://travis-ci.org/pantsbuild/pants/jobs/62120623

Testing Done:
Relevant test passes locally in otherwise unmodified repo.
Relevant test fails with INTERNALERROR when test is deliberately broken before this change.
Relevant test fails with useful backtrace when deliberately broken after this change.

CI pending: https://travis-ci.org/pantsbuild/pants/builds/62150849

Bugs closed: 1525

Reviewed at https://rbcommons.com/s/twitter/r/2201/