sanity check on generated cobertura xml report
One of our users was getting zero hits reported in the cobertura xml
report, which is used by the cobertura jenkins plugin. His generated
html was fine. We couldn't track down the error but this patch adds a
test for that condition. We needed some kind of test on the generated
cobertura reports anyway.