Fixup OSX CI breaks.
Kill now-dead `-e` and `-t` flags in one of the OSX CI shards.

Fixup Distribution to discover its true home dir even if initial
discovery is via the jre `java` executable and add unit tests to
exercise this.  Also fixup both binary and lib_path discovery to
check both JDK and JRE bin and lib paths when the distribution is a JDK.