add the goal "attributes"
We need a way of determining what platform a target is meant to run on,
or what language the target is for. We have scripts that behave
differently for a given target. If it is python, we parse the output one
way, if it is jvm, we parse it another. Similarly, depending on the
platform, we may retry a failing goal differently.