Extract the android plugin as an sdist.
This change extracts the existing android backend as its own
distribution to help with
https://github.com/pantsbuild/pants/issues/1552, although ~all backends
will be extracted like this; so this also serves to lay the groundowrk
for those other extractions.

In order to support the release.sh and ci.sh scripts, some re-work of
source root registrations was needed.

In addition, the existing `--bootstrap-buildfiles` option was
in-effective in global scope since there is a 'bootstrap' goal that
claims the namespace.  Instead of attempting to fix this subtle problem
with namespace shadowing, I lifted up a subsystem for reading bootstrap
build files and reverted to the old namespace of goals.  This solves the
shadowing problem, isolates the bootstrap code to be killed better, and
presents slightly less thrash to users since the
`[DEFAULT]bootstrap_buildfiles` change has not been released yet.