Refactoring ProtobufGen, JaxbGen, and WireGen to use Patrick's SimpleCodegenTask base class.
As per John Sirois's recommendation, the python generation code was removed from ProtobufGen,
because it apparently isn't used and muddles the purpose of the JavaProtobufLibrary target.

There are some TODO's for integrating isolated code-generation strategies. I have this
functionality mostly working in a separate branch, which I hope to put up for review
separately once this patch is landed.

Fixed jar-publish bug.