Shade zinc.
Zinc can run javac in-process and when it does so it uses no
classpath isolation mechanisms and so can cause annotation processors
run under javac to come into conflict with its own classpath.  This
change shades zinc to avoid the problem altogether.

In so doing, a few changes were needed:
+ Turn on the local artifact cache for bootstrap in the
  pantsbuild/pants repo.  Although the junit-runner shade time is ~5s
  the zinc shade time is ~30s making the need for this acute.
+ Refactor ScalaCompile to explicitly ask for classpaths for
  sbt-interface and compiler-interface.  This work inlines ZincUtils
  which was slated for this.

Along the way some unused lazy tool classpath methods were killed and
a new tool_jar helper was added for tool classpaths that are
expected to have exactly one entry.