Allow jvm tool tests to bootstrap from the artifact cache.
This makes those tests quite a bit faster. For example,

./pants test tests/python/pants_test/backend/jvm/tasks:junit_run

Runs in 35 seconds with no cached junit jar, but in under 5 seconds with.

This was also a useful exercise in reasoning about test isolation:
To use cached jars they must have the same path under .pants.d, and
in our tests those paths included a randomly named tmpdir and
a randomly named options scope. Switching those to stable names
forced me to think about whether that was OK, and my conclusion was
that it is, which is a good sign for our test isolation.