Remove unused imports from python files
This was done by running Intellij's Optimize Imports tool over the
python files. I then checked the BUILD files against the diff manually.

I reverted a couple of style conflicts with isort (intellij moves
'import sys' to its own section) but other than that this was easy
and took 15-20 mins. The only tedious part was that many BUILD files
use "dependency bags" which meant I often had to grep the entire
module to see if a dep could be removed from the bag or not.