Invoke jvm doc tools via java.
This removes dependence on the `javadoc` and `scaladoc` binaries which
immediately opens up javadoc and scaladoc testing on Travis-CI.

For `javadoc` this is not so important today, since the new invocation
scheme requires a jdk for `tools.jar` which necessarily means `javadoc`
is present.  That said, it does set the stage for invoking javadoc
using the correct JDK based on the repo-wide or per-target JDK
constraints.