Avoid ivy cache thrash due to ivydata updates.
The ivydata files' effect on resolution time is in the noise, but
they are re-timestamped in a comment by ivy on each run and lead to
cache thrash in travis-ci.  Kill these files before the cache check to
avoid un-needed cache re-packing and re-upload which is a ~100s
operation that occurs at the end of the build on 9 or our 10 ci shards.