Avoid ivy cache thrash due to ivydata updates.
The `ivydata-*.properties` and root-level `*.{properties,xml}` files'
effect on resolution time is in the noise, but they are re-timestamped
in internal comments and fields by ivy on each run and this leads to
cache thrash in travis-ci. Kill these files before the cache check to
avoid un-needed cache re-packing and re-upload which is a ~100s
operation that occurs at the end of the build on 9 of our 10 ci
shards.

Testing Done:
Ran `travis lint`, which fails with:
```console
$ travis lint
Warnings for .travis.yml:
[x] unexpected key before_cache, dropping
```

However, this is a newish documented feature:
  https://github.com/travis-ci/docs-travis-ci-com/commit/088be0f8de1c45135bcfa6d3b48ee0a561448716

As a result, giving this a whirl since the savings should be large if
the feature does actually exist.

CI went green here:
  https://travis-ci.org/pantsbuild/pants/builds/65978345

Bugs closed: 1650

Reviewed at https://rbcommons.com/s/twitter/r/2333/