Fix bootstrap local cache.
c960c9c3 introduced a finer scope to the `bootstrap` local cache turn-on
in pants.ini and the scope was in-error (used `_`s instead of `-`s).
Instead of fixing the underscores, just revert the change since it was
correct to cache bootstrap all bootstrap tasks in the first place.

Testing Done:
Noticed the 'Caching artifacts for 1 target.' messages gone on master,
with the revert the caching is back.

CI is away here:
  https://travis-ci.org/pantsbuild/pants/builds/66085108

Bugs closed: 1653

Reviewed at https://rbcommons.com/s/twitter/r/2336/