Don't use soft excludes for mapjars
soft excludes was only intended to change how the global resolve works so that subgraphs containing dependency trees with overlapping excludes could still resolve and compile. It doesn't make sense to use it for map jars as well since mapjars is already isolated by target

Testing Done:
built a bundle for a target with soft-excludes on and saw that it didn't affect the exclusion of excluded jars in the resulting lib dir. Added a test to cover the regression.

Bugs closed: 1654

Reviewed at https://rbcommons.com/s/twitter/r/2330/