Add INDEX.LIST to as a Skip JarRule when creating a fat jar
Testing Done:
The INDEX.LIST files are usually created with `jar i` as some kind of optimizing step after a .jar file is built.

I found a problem when trying to add `Class-Path:` to a manifest file when building a fat jar.  The root cause was an INDEX.LIST file from one of the dependent jars that got combined together.  When this file is present, `Class-Path:` seems to be ignored.

The INDEX.LIST files are scrubbed out by default with the maven shade plugin:

https://github.com/intelie/maven-shade-plugin/blob/71b5895028f9c9ca2730b45d1117d8e6d3372a3e/src/main/java/org/apache/maven/plugins/shade/DefaultShader.java#L137

Bugs closed: 1665

Reviewed at https://rbcommons.com/s/twitter/r/2342/