Ensure caliper is shaded in bench, add bench desc, use RUN so that output is printed
caliper depends on guava 15 which is incompatible with 16 and 18 due to method deprecations, adding main to the jvm tool registration causes it to be shaded, which means the older guava won't interfer with code under test that depends on newer guava versions.

I also added a description and set the WorkUnit.RUN label so that the caliper output will be printed to the console by default.