Fixed a bug where codegen would crash due to a missing flag.
Specifically, the error message (or warning, depending on the flag) printed if a
target generates the same sources as one of its parents.

Also fixed a bug where an exception message was in the wrong namespace, which would
have crashed nastily if it was ever thrown.

Added/updated simple_codegen_task test-cases to cover these things.