Ensure caliper is shaded in bench, add bench desc, use RUN so that output is printed
caliper depends on guava 11 which is incompatible with guavas above 15 due to deprecated method removals, adding main to the jvm tool registration causes it to be shaded, which means the older guava won't interfer with code under test that depends on newer guava versions.

I also added a description and set the WorkUnit.RUN label so that the caliper output will be printed to the console by default.

Testing Done:
tested against a target with a dependency on guava 16, which failed to run without shading. CI in PR

Bugs closed: 1674

Reviewed at https://rbcommons.com/s/twitter/r/2353/