Add a CONTRIBUTORS maintenance script.
This includes a .mailmap for regularizing contributor ids as well as
handling of the portion of the zinc subtree history before-pants.

Testing Done:
Used the script to re-generate the CONTRIBUTORS.md as seen in this
change.  Also ran the the script to get the contribution roster since
the 0.0.33 release and verified it was accurate by hand:
```console
$ ./build-support/bin/contributors.sh -s release_0.0.33
      2 Eric Ayers
      1 Nick Howard
      1 Kris Wilson
      1 John Sirois
      1 Benjy Weinberger
```

CI went green here:
  https://travis-ci.org/pantsbuild/pants/builds/67089966

Bugs closed: 1695

Reviewed at https://rbcommons.com/s/twitter/r/2377/