Promote PANTS_DEV=1 to the only ./pants mode.
When hacking on pants it only makes sense to run pants from live
sources.  Remove the mode that generated a pants.pex and ran from it
since it only led to confusing breaks and offered no benefit whatsoever.

The CI script still builds a local pex to prove pants can bootstrap
itself as a pex binary, but it does that directly anyhow.