A JVM subsystem.
This is primarily for tasks whose job is to run a JVM, such as
JunitRun, ScalaRepl, JvmRun etc. Basically, anything that extends
JvmTask.

This is not for tasks that happen to use some JVM-based tool to
do their work, i.e., tasks that mix in JvmToolTaskMixin.  A separate
change will subsystem-ize those tools.

The distinction is necessary because, for example, with the former
it makes sense to set the program args via an option, whereas
with the latter the program args are determined by the task.