Make sure all Task mixins are on the left.
Note that tasks that extend NailgunTask cannot have JvmToolTaskMixin on
the left, because NailgunTask is already a JvmToolTaskMixin, and this leads
to a "Cannot create a consistent method resolution order (MRO)" error.
But the solution is simply not to explicitly mix in JvmToolTaskMixin,
and rely on inheriting it from NailgunTask.

Testing Done:
CI passes: https://travis-ci.org/pantsbuild/pants/builds/68404283.

Reviewed at https://rbcommons.com/s/twitter/r/2421/