Fix test_execute_fail.
This fixes test_execute, but not the underlying global state used in
DummyTask.  Its that state, that appears to have supported a successful
run under certain sharding / test method ordering, although I was not
able to positively identify a sequence of calls that would lead to
test_execute ever passing.  In particular, it had 2 local errors that
look systemic:
1. The raises regex did not escape the parens.
   Before this fix, the regex only could match 'Failed to generate
   targets' - with s being in an unused capture group.
2. The assertRaisesRegexp callable argument was not a callable.
   The _result_ of task.execute() was passed instead of the task.execute
   method.

These system errors were in addition to the lack of setup of the global
strategy explicitly.  That strategy is needed to ensure DummyGen's
execute_codegen is called leading to the raise.  Using the isolated
strategy (which was the default) with no targets sort circuits the
execute_codegen call, failing to ever engage the `should_fail` forced
raise.