Fix test bugs introduced in b22024802.
That change turned on artifact caching globally in tests.
The outcome was that many tests were actually not doing anything.
Instead, they were reading from the real artifact cache.

So, for example, the antlr gen test wasn't running antlr - it was
finding the expected outputs in the cache...

This commit removes the code that turned universal caching on,
and then re-enables caching just for tool bootstrapping, for test
performance.

This fixed the weird issue with antlr gen, so that large comment
is gone too.

Incidentally, this commit also changes the way we register the
source root in the antlr gen test. Instead of eval'ing it in a
BUILD file it just registers it directly, as other tests do.