Switch almost all python tasks over to use cached chroots.
Now that we're on pex 1.0+, this works. And indeed, the effect is dramatic:
On my laptop running all the unittests went from ~5:00 mins to around 1:45 mins.

Note that binary_create doesn't get any benefit from this yet, because the
pex_info contains timestamps, which makes the cache key miss every time.

We can think about a fix for this, and then get rid of temporary_chroot()
entirely.

Also fixes a few random nits that I noticed on the way.