Add android_library target and unpack_library task
This review is part 1 of splitting review #2040. This adds
android_library and android_dependency targets as well as an
unpack_libraries task. This is not integrated with the rest
of the android backend, that will be part 2.

An android_dependency can be a jar or aar library artifact, in
a repo or from the SDK. Those artifacts have to be unpacked and then
repacked into the final android application.

Each library is unpacked exactly once. The includes/excludes
will be handled in the repacking. That keeps us from having to repeatedly
unpack common dependencies. They all have a common base so this
treats the unpacked source as this common starting point.

If the library is an aar file, it needs to add targets to the
graph based on the aar's contents, as seen in UnpackLibraries.

The follow up will integrate the tasks/targets into the other tasks
and finish up adding the testing harness. If you'd like to see how
it runs or what the BUILD files will look like, you can check #2040.

references #1390

Testing Done:
Passes local ci.sh and travis.

Bugs closed: 1800

Reviewed at https://rbcommons.com/s/twitter/r/2467/