Fixup SetupPy handling of exported thrift.
SetupPy was attempting to create a PythonThriftBuilder with an invalid
set of constructor args.  This change adds a test that exercises this
codepath in addition to the path for exported antlr libraries.

In the course of creating a testable PythonThriftBuilder, extract
BinaryUtil.Factory and ThriftBinary.Factory subsystems to both
encapsulate options and allow for directly constructable BinaryUtil and
ThriftBinary objects.

Some further file move cleanups are marked TODO in favor of keeping the
scope of this review from ballooning farther than it already has.

Testing Done:
CI went green here:
  https://travis-ci.org/pantsbuild/pants/builds/72214777

Bugs closed: 1805, 1847

Reviewed at https://rbcommons.com/s/twitter/r/2511/