Invalidate tasks based on BinaryUtil.version.
The assumption is that binaries are uniquely identified by their version
and name. This is at least true of pants-supported bintray binaries for
protoc and ragel.

Additionally invalidation is arranged for protoc plugins based on name.
This is not ideal - it allows for changing a plugin without chaging its
name - but its strictly better than not invalidating at all for changing
plugin lists.