Refactor and simplify aapt_gen
Switching to just using the target_sdk of the android_binary
left a lot of opportunity for simplification that I didn't
pursue in the first pass.

Also, be more specific about what aapt_files are unique.
Previously we were invalidating targets if the R.java's
relative path existed, but that didn't account for
if the library needed to be processed by multiple
sdks. This uses the containing outdir, which has the sdk number
in the path.