Moved logic which validates jvm platform dependencies.
The logic which checked to make sure no targets depended on other
targets with newer jdks was previously in jvm_compile, which caused
it to be re-ran for every single type of jvm compilation. It also
failed fast and had no option to simply warn and continue, which
made it hard and frustrating to debug.

JvmPlatformValidate is a new task whose sole purpose is performing
this validation. It runs fairly early in the build lifecycle, which
prevents pants from doing too much work before erroring on a doomed
run.

It also allows error message to be disabled, or demoted to a warning.

Also included is a jvm platform debugging resource which I developed
(along with the previous change) to help me sort out many
jvm platform dependency problems in square's internal java repo.

This is done in the JvmPlatformExplain console task, which performs
a more detailed analysis of the target level constraints imposed on
targets by their dependencies and dependees (a target cannot have a
higher -target than its dependees, or a lower -target than its
dependencies).

Existing tests modified, additional unit tests added in
test_jvm_platform_analysis.py.