add support for android compilation with scrooge
Problem

New version of scrooge 3.20 support compiling thrift file for android
(which limit the number of methods created in each class).

Solution

Upgraded Scrooge to 3.20 and fix old unit test to work with this new
version. Modified the scrooge_gen.py so that it supports this new 'language'.
Added some unit test to verify the new functionality.