Introduce a classpath specific product to encapsulate excludes behavior
I'm working on improving excludes behavior. As part of that, I've started encapsulating excludes / classpath generation so that we can more easily change the internal representation of classpath elements and what excludes are represented as during application. This first step towards that just moves exclude handling into a product along with classpaths out of ClasspathUtils.

With this change, the responsibilities are laid out this way:

ClasspathUtils is responsible for

* filtering classpath tuples by conf
* validating paths exist within the buildroot

ClasspathProducts is responsible for

* precomputing excludes and classpath graphs
* applying excludes as classpath tuple lists are constructed
* constructing classpaths based on the precomputed graphs

Also, if a target provides a jar, it adds an exclude for that jar to the excludes used for creating classpaths

Follow on work:

* I'm working towards applying excludes only to the child dependencies of the targets that define them rather than applying them to the full closure of the target whose classpath is being created. That way they'll work as documented. They'll also be easier for users to reason about because they won't have to consider how excludes they introduce may interact with excludes from a separate dependency of a dependee.

* Move towards https://docs.google.com/document/d/1sEMXUmj7v-YCBZ_wHLpCFjkHOeWjsc1NR1hRIJ9uCZ8/edit# by internally representing the classpath elements as python objects that look like those described in the doc

Testing Done:
Passing PR https://github.com/pantsbuild/pants/pull/1637

I'm also doing some internal verification.

Bugs closed: 1637

Reviewed at https://rbcommons.com/s/twitter/r/2541/