A stats db.
Stores timing stats in a structured way, instead of as json.
Future changes will do interesting visualizations and analytics
on this data.

Tightened up some uses of RunInfo, as previously we were inadvertently
overwriting the id with None. Also ensures that ids are unique, by
appending a uuid to them (but we still prefix them with a timestamp
so that they sort by time).