Add an affordance for 1 flag implying another.
This adds an is_flagged helper to the OptionValueContainer
(get_options()), and the JVM subsystem leverages it to turn on debugging
whenever a user explicitly flags a custom debug port or custom debug
args on the command line.

The change also introduces a subsystem test utility and refactors option
value container faking.