Fix some more missing globs in build-file-rev mode.
Some of these globs were caused because resources eagerly converted
the sources field to a list before its filespec had a chance to be
captured.  Others were because has_sources was called on some targets,
which checks for any on-filesystem source files; in globs mode, we
only care about globs, not actual files, so we don't care if a target
has sources.

I believe this is the lat of the build-file-rev related changes for now.