Speeding up jvm-platform-validate step.
Using self.invalidated mechanism, targets are only re-evaluated if
they (1) they failed previously, (2) their dependencies have changed,
or (3) their platform -target version has changed.

Additionally, pruned down the list of dependencies that must be
considered for each target. Instead of evaluating all transitive
JvmTarget dependencies, only consider the "direct" JvmTarget
dependencies. "Direct" is in quotes because I'm defining a JvmTarget
'A' as "directly" depending on a JvmTarget 'B' iff there exists a
path in the dependency graph from 'A' to 'B', and there are no
internal vertices in the path that are JvmTargets.

This behavior is documented in the jvm_dependency_map property.