Elevate the pants version to a 1st class option.
Currently `pants_version` is a config option relied on by pants tooling
but not formalized as a registered option in the pants codebase.  This
change lifts all forms of pants version (`-V`, `--version` and the new
`--pants-version`) to be a global option.  The option is configured to
work as both a 0-arg command-line flag to trigger version help printing
as well as an option to use for setting the desired pants_version in
pants.ini (or via an the PANTS_VERSION env var).  The pantsbuild/setup
script and other proprietary pants bootstrap scripts use this option to
bootstrap pants to the requested version today and in the future pants
itself will use the version in this way.