Re-Add debug option to benchmark run task, complain on no jvm targets, add test
The jvm subsystem options change broke the bench task. This adds the debug option back in as a task option since it means something slightly different in the bench task.

It also causes the task to fail if there are no jvm targets reachable from the target roots.

I found it a little tricky to write a test that fully exercise the task in a unit test. I think an integration test would be a good idea for this, but I wanted to fix the regression first.

Testing Done:
Wrote a short test, saw it fail with the missing debug option field. Added a registration for it and saw it pass. CI in https://github.com/pantsbuild/pants/pull/1974

Bugs closed: 1931, 1974

Reviewed at https://rbcommons.com/s/twitter/r/2619/