Elevate the pants version to a 1st class option.
Currently `pants_version` is a config option relied on by pants tooling
but not formalized as a registered option in the pants codebase.  This
change lifts all forms of pants version (`-V`, `--version` and the new
`--pants-version`) to be a global option.  The option is configured to
work as both a 0-arg command-line flag to trigger version help printing
as well as an option to use for setting the desired pants_version in
pants.ini (or via an the PANTS_VERSION env var).  The pantsbuild/setup
script and other proprietary pants bootstrap scripts use this option to
bootstrap pants to the requested version today and in the future pants
itself will use the version in this way.

Testing Done:
Verified a 4x speed improvement in `./pants --version`! - heh.

Verified pants_version is picked up as a pants.ini option like so:
```diff
$ git diff -U1 pants.ini
diff --git a/pants.ini b/pants.ini
index a557fbf..8b2eeea 100644
--- a/pants.ini
+++ b/pants.ini
@@ -11,2 +11,4 @@
 [DEFAULT]
+pants_version: jake
+
 print_exception_stacktrace: True
```
```console
$ ./pants -V
0.0.41
$ ./pants -h | grep version
-V=<str>, --pants-version=<str>, --version=<str> (default: jake)
    Prints pants' version number and exits.
```

Note that the real version of pants is printed when -V is requested
on the CLI, but the configured pants_version print as the default in
help output.  Once pantsbuild/setup (next change) and, in the future,
pants itself uses pants_version to bootstrap the proper pants version
this will always align as expected.

CI is away here:
  https://travis-ci.org/pantsbuild/pants/builds/75679069

Bugs closed: 1983

Reviewed at https://rbcommons.com/s/twitter/r/2627/