Support for running junit tests with different jvm versions.
Junit tests now run using a java Distribution found to match the
test_platform.target_level of the test target. By default any jvm
whose version is *at least* the target_level is used. Using jvms
which are of the exact same version of the junit_tests can be forced
via the --strict-jvm-version flag.

If the test_platform isn't specified, it is assumed to be the same
as the platform (inherited from jvm_target).